# Optimize the performance of Marshal/Unmarshal for WKB