It turns out that using `binary.Read` and `binary.Write` is quite slow. Significant performance gains can be achieved by using `byteOrder.Uint64` and `byteOrder.PutUint64`.

Here are some benchmarks:

Before my change:
```
BenchmarkUnmarshal-4   	    5000       	    230946 ns/op       	  133568 B/op  	    2290 allocs/op
BenchmarkMarshal-4     	   10000       	    231463 ns/op       	   87008 B/op  	    1753 allocs/op
```
After my change:
```
BenchmarkUnmarshal-4   	   10000       	    140244 ns/op       	  120400 B/op  	    1542 allocs/op
BenchmarkMarshal-4     	   10000       	    117554 ns/op       	   73776 B/op  	    1005 allocs/op
```