@twpayne 2016-09-20
Fantastic improvement, many thanks!